petties-agent-serivce/requirements.txt
# ============================================
# PETTIES AI AGENT SERVICE - Python Dependencies
# Core dependencies for production (Render free tier compatible)
# Use requirements-full.txt for local development with all features
# ============================================

# ==================== FastAPI Framework ====================
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
pydantic>=2.10.3
pydantic-settings>=2.6.1

# ==================== Async & WebSocket ====================
websockets>=14.1
python-multipart>=0.0.19

# ==================== AI & Agent Orchestration ====================
langchain>=0.3.13
langchain-community>=0.3.13
langchain-core>=0.3.28
langchain-openai>=0.2.0
langchain-ollama>=0.1.9
langgraph>=0.2.60
llama-index>=0.11.20

# ==================== LLM Clients ====================
openai>=1.54.4
ollama>=0.4.5  # For Ollama Cloud support

# ==================== MCP Framework ====================
fastmcp>=2.3.3

# ==================== Database ====================
sqlalchemy[asyncio]>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0

# ==================== Vector Database ====================
qdrant-client>=1.12.1

# ==================== Document Processing ====================
pypdf2>=3.0.1
python-docx>=1.1.2

# ==================== Web Search ====================
duckduckgo-search>=7.2.1
beautifulsoup4>=4.12.3

# ==================== HTTP Clients ====================
httpx>=0.27.2
aiohttp>=3.11.2
requests>=2.32.3

# ==================== Data Processing ====================
pandas>=2.2.3
numpy>=1.26.4,<2.0.0

# ==================== Authentication ====================
python-jose[cryptography]>=3.3.0  # JWT token handling

# ==================== Utilities ====================
python-dotenv>=1.0.1
pyyaml>=6.0.2
tenacity>=9.0.0

# ==================== Monitoring & Logging ====================
prometheus-client>=0.21.1
loguru>=0.7.3
